create HomeScreen

+++ app/src/main/java/com/moonwalkin/numbertesttask/data/network/NumberService.kt
@@ -5,7 +5,7 @@ import retrofit2.http.Path
 
 interface NumberService {
     @GET("/{number}?json")
-    suspend fun getNumberInfo(@Path("number") number: Int): NumberInfoDto
+    suspend fun getNumberInfo(@Path("number") number: Long): NumberInfoDto
 
     @GET("/random/math?json")
     suspend fun getRandomNumberInfo(): NumberInfoDto
